backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2==2.9.9
python-dotenv==1.0.1
yfinance==0.2.36
pandas==2.2.0
pandas_datareader==0.10.0
requests==2.31.0
redis==5.0.1
fastapi-cache2==0.2.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pytz==2024.1 